feat: Define VM's public_network ip and device
Create two new variables `bootstrap_public_ip` and
`bootstrap_public_dev` to respectively define the IP address of the VM
and the network device to use on the host.

Even if the variables are used in the Vagrantfile, declare these in the
Ansible inventory so there is a single configuration file to edit in
order to bootstrap the Cobbler server.