feat: Allow to define an alternate nameserver
The Vagrant box `generic/debian11` comes pre-configured with public DNS
servers. Sometimes it might be convenient to use a corporate DNS instead
to access internal resources (e.g. an internal mirror of a distro
repository).

Add an optional variable `bootstrap_public_nameserver` that allows to
define an alternate nameserver, both in the interface configuration and
in the file /etc/resolv.conf (to avoid a restart of the network, which
does not behave well when configuring a new Vagrant guest).